refactor(architecture): refactor auth
