+++ router/auth/swagger/deleteUser.swagger.js
@@ -1,0 +1,14 @@
+/**
+ * @swagger
+ * /delete-user:
+ *   delete:
+ *     summary: Delete the currently authenticated user
+ *     tags: [Auth]
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: User deleted successfully
+ *       401:
+ *         description: Unauthorized
+ */
